+++ telemetry/metrics_test.go
@@ -15,12 +15,29 @@ func TestMetricPayload(t *testing.T) {
 	// attributes correctly marshals into JSON.
 	now := time.Date(2014, time.November, 28, 1, 1, 0, 0, time.UTC)
 	h, _ := NewHarvester(ConfigCommonAttributes(map[string]interface{}{"zop": "zup"}), configTesting)
-	// Use a single metric to avoid sorting.
 	h.RecordMetric(Gauge{
 		Name:       "metric",
 		Attributes: map[string]interface{}{"zip": "zap"},
 		Timestamp:  now,
 		Value:      1.0,
+	})
+	h.RecordMetric(Summary{
+		Name:       "summary-metric-nan-min",
+		Attributes: map[string]interface{}{"zip": "zap"},
+		Timestamp:  now,
+		Count:      4.0,
+		Sum:        1.0,
+		Min:        math.NaN(),
+		Max:        3.0,
+	})
+	h.RecordMetric(Summary{
+		Name:       "summary-metric-nan-max",
+		Attributes: map[string]interface{}{"zip": "zap"},
+		Timestamp:  now,
+		Count:      4.0,
+		Sum:        1.0,
+		Min:        10,
+		Max:        math.NaN(),
 	})
 	h.lastHarvest = now
 	end := h.lastHarvest.Add(5 * time.Second)
@@ -37,7 +54,9 @@ func TestMetricPayload(t *testing.T) {
 			"attributes":{"zop":"zup"}
 		},
 		"metrics":[
-			{"name":"metric","type":"gauge","value":1,"timestamp":1417136460000,"attributes":{"zip":"zap"}}
+			{"name":"metric","type":"gauge","value":1,"timestamp":1417136460000,"attributes":{"zip":"zap"}},
+			{"name":"summary-metric-nan-min","type":"summary","value":{"sum":1,"count":4,"min":null,"max":3},"timestamp":1417136460000,"attributes":{"zip":"zap"}},
+			{"name":"summary-metric-nan-max","type":"summary","value":{"sum":1,"count":4,"min":10,"max":null},"timestamp":1417136460000,"attributes":{"zip":"zap"}}
 		]
 	}]`
 	compactExpect := compactJSONString(expect)
@@ -151,10 +170,15 @@ func TestValidateGauge(t *testing.T) {
 }
 
 func TestValidateSummary(t *testing.T) {
-	want := map[string]interface{}{
+	expectNaNErr := map[string]interface{}{
 		"message": "invalid summary field",
 		"name":    "my-summary",
 		"err":     errFloatNaN.Error(),
+	}
+	expectInfErr := map[string]interface{}{
+		"message": "invalid summary field",
+		"name":    "my-summary",
+		"err":     errFloatInfinity.Error(),
 	}
 	testcases := []struct {
 		m      Summary
@@ -165,20 +189,24 @@ func TestValidateSummary(t *testing.T) {
 			fields: nil,
 		},
 		{
+			m:      Summary{Name: "my-summary", Count: 1.0, Sum: 2.0, Min: math.NaN(), Max: math.NaN()},
+			fields: nil,
+		},
+		{
 			m:      Summary{Name: "my-summary", Count: math.NaN(), Sum: 2.0, Min: 3.0, Max: 4.0},
-			fields: want,
+			fields: expectNaNErr,
 		},
 		{
 			m:      Summary{Name: "my-summary", Count: 1.0, Sum: math.NaN(), Min: 3.0, Max: 4.0},
-			fields: want,
+			fields: expectNaNErr,
 		},
 		{
-			m:      Summary{Name: "my-summary", Count: 1.0, Sum: 2.0, Min: math.NaN(), Max: 4.0},
-			fields: want,
+			m:      Summary{Name: "my-summary", Count: 1.0, Sum: 2.0, Min: math.Inf(-3), Max: 4.0},
+			fields: expectInfErr,
 		},
 		{
-			m:      Summary{Name: "my-summary", Count: 1.0, Sum: 2.0, Min: 3.0, Max: math.NaN()},
-			fields: want,
+			m:      Summary{Name: "my-summary", Count: 1.0, Sum: 2.0, Min: 3.0, Max: math.Inf(3)},
+			fields: expectInfErr,
 		},
 	}
 	for idx, tc := range testcases {